Add TrajOptFreespacePlanner
This planner is intended to provide an easy to use interface to TrajOpt for freespace planning. It is made to take a start and end point and automate the generation of the TrajOpt problem. It does not expose all of the configuration available in TrajOpt, but is made in hopes of lowering the barrier to entry for using TrajOpt to solve simple problems.