Fix the way in which Eigen is included
Currently, Eigen is added via `target_include_directories` as a PUBLIC
or INTERFACE tag. Unfortunately, this makes a built tesseract
non-reusable on a system where Eigen is installed in a different folder.
For example, Tesseract built on a system where Eigen was downloaded in
the ROS workspace is not usable on a system where Eigen is installed in
the default system directory.