A few fixes that were needed for Windows (#708)
* Make HACDConvexDecomposition library optional

Bullet extras are not easily obtained on Windows. If found, build library, otherwise ignore. Also the plain ConvexDecomposition library is looked for but never used and so removed entirely.

* Check if Bullet CMake variables are using absolute paths

For some reasons, the vcpkg ported version changes the config file to
use absolute paths instead of relative to BULLET_ROOT_DIR

* Add include for std::string

Co-authored-by: Levi Armstrong <levi.armstrong@gmail.com>